Move circuit.fuse()
